Update WOLtoDHCP.py
Created two functions for extracting the name and mac address. Rather then using a static index to use for the substring, this version will automatically check the index of where the tags are and pull the text in between the open/closing tags. This will allow the name/mac to be whatever length as well. Will also allow me to extract other info down the road using similar functon structure